requirements.txt
crewai>=0.28.0
anthropic>=0.5.0
requests>=2.31.0
python-dotenv>=1.0.0
asyncio>=3.4.3
pydantic>=2.0.0
loguru>=0.7.0 